chore(runtime): token-diet phase 2-A — split companion/outbox.rs
Pure code movement, behavior unchanged. Splits the 2265-line
mur-agent-runtime/src/companion/outbox.rs into a folder so every file
satisfies the project's 800-line limit:

  outbox/
    mod.rs       (726)  doc + types + Outbox struct + run_tick orchestrator
    generate.rs  (146)  steps 8+9 — generate_with_lint + prompt-seed substitute
    i18n.rs       (75)  step 10 — handle_i18n + I18nResult
    deliver.rs    (80)  steps 11+12 — deliver_and_finalise
    tests/
      mod.rs     (240)  shared FakeNotifier / fixtures / make_outbox helpers
      rhythm.rs  (172)  M5.3 — proactive disabled / cap / morning / rollover
      lint.rs    (225)  M5.4 — generate+lint regenerate / persistent / notifier
      i18n.rs    (407)  M5.5 — translate / 429 backoff / 4-failure / rate-limit
      dismiss.rs (239)  M5.6 — passive-dismiss sweep

Internal step helpers move into per-step `impl Outbox<R>` blocks with
`pub(super)` visibility so `run_tick` (in mod.rs) can call them.
`GenerateResult` / `I18nResult` enums move alongside the steps that
return them. `RETRY_BACKOFF_SECS` / `backoff_for_attempt` stay in
mod.rs because both run_tick and the i18n step use them.

network_audit.rs updated to embed the four production outbox files
individually and to count outbox submodules (excluding `mod tests;`)
when verifying COMPANION_FILES coverage.

cargo build --workspace + cargo test -p mur-agent-runtime pass.
cargo clippy -p mur-agent-runtime --lib clean. (Pre-existing clippy
warnings in mur-common/tests/companion_enums.rs and
mur-agent-runtime/tests/scheduler_integration.rs are unchanged on main
and out of scope for this code-movement PR.)
